move reload handler logics to local storage service
