+++ src/app/booking/booking.component.ts
@@ -1,56 +1,19 @@
-import { Component, HostListener, OnDestroy, OnInit } from '@angular/core';
-import { BookingService } from '../core/services/booking.service';
-import { FullBookingInfo } from '../shared/models/booking';
-import { LOCAL_STORAGE_BOOKING_KEY } from '../shared/constants/string-constants';
+import { Component, HostListener, OnDestroy } from '@angular/core';
+import { LocalStorageService } from '../core/services/local-storage.service';
 
 @Component({
   selector: 'app-booking-page',
   templateUrl: './booking.component.html',
 })
-export class BookingComponent implements OnDestroy, OnInit {
-  public constructor(private bookingService: BookingService) {}
-
-  public ngOnInit(): void {
-    this.getBookingFromLocalStorage();
-  }
+export class BookingComponent implements OnDestroy {
+  public constructor(private localStorageService: LocalStorageService) {}
 
   @HostListener('window:beforeunload', ['$event'])
   public beforeUnloadHandler(): void {
-    this.saveBookingToLocalStorage();
-  }
-
-  @HostListener('window:load', ['$event'])
-  public onloadHandler(): void {
-    this.getBookingFromLocalStorage();
-  }
-
-  private saveBookingToLocalStorage(): void {
-    const booking = this.bookingService.getCurrentBookingInfo();
-    const selected = this.bookingService.getCurrentSelectedFlights();
-    const passengers = this.bookingService.passengersInfo;
-    const bookingInfo = {
-      booking,
-      selected,
-      passengers,
-    };
-    localStorage.setItem(
-      LOCAL_STORAGE_BOOKING_KEY,
-      JSON.stringify(bookingInfo)
-    );
-  }
-
-  private getBookingFromLocalStorage(): void {
-    const ls = localStorage.getItem(LOCAL_STORAGE_BOOKING_KEY);
-    if (ls) {
-      const bookingData: FullBookingInfo = JSON.parse(ls);
-      this.bookingService.setBookingInfo(bookingData.booking);
-      this.bookingService.addForwardFlight(bookingData.selected.forwardFlight);
-      this.bookingService.addReturnFlight(bookingData.selected.returnFlight);
-      this.bookingService.passengersInfo = bookingData.passengers;
-    }
+    this.localStorageService.saveBookingToLocalStorage();
   }
 
   public ngOnDestroy(): void {
-    localStorage.removeItem(LOCAL_STORAGE_BOOKING_KEY);
+    this.localStorageService.deleteBookingFromLocalStorage();
   }
 }
